mcp-server/pyproject.toml
[project]
name = "text2sql-mcp"
version = "0.1.0"
description = "MCP Server for Text 2 SQL"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.0",
    "uvicorn>=0.24.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "fastembed>=0.2.0",  # ONNX Runtime-based embeddings
    "numpy>=1.24.0",     # Required by fastembed
    "mlflow>=2.11.0",

    "neo4j>=5.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "openai>=1.0.0",
    "hnswlib>=0.8.0",  # High-perf ANN for vector search
    "spacy>=3.7.0",    # NLP for linguistic canonicalization
    "sqlglot>=25.0.0",  # SQL AST parsing and validation
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "hnswlib>=0.8.0",  # Needed for HNSWIndex tests
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_server*"]
